[pivot] Solve systems with vector variables
The rasm project will allow users to programatically create drawings by giving
a set of components and separately imposing spacial constraints for those
components. The spacial constraints will be in the form of systems of linear
equations where each variable actually denotes a 2d vector. To support this
interface this change introduces a new engine derived from the linear engine
which reduces equations of 2d vector variables to equations of scalar variables.