[xonshlib] Support for using development versions of my libraries
I sometimes like to make modifications to multiple python projects at the same
time and test their integration. This change lets me easily modify add and
remove source directories of python projects to the PYTHONPATH env variable. It
may also be nice to have an indicator in my prompt to tell me if I am in
"devmode" for any of my projects.