[xonshlib] Extension to aggreate to-do items
There are a number of places where I track things that need to be done:

* bookmarks for pages I need to visit or process
* e-mails I need to process or that need a response
* github issues for coding work
* errands and other reminders in my notes
* TODO comments in code blocks

While it makes sense to interact with each of these items in its orginal
application I would like to have a single high-level view of all of my items so
that I can know if there is any uncategorized or high-priority work remaining
and which of my categories are most the most neglected. This change introduces
a pluggable extendable to-do item aggregator in xonshlib as well as extensions
for Github, GMail, and Chrome bookmarks (the other two extensions still
remain). It also plugs those components into my xonsh dotfile and adds a
'tasks' alias for running the aggregation.

closes #17