[bin] Add git pr command
I have a very common flow of pushing a branch to my fork, going to the repo on
github, and creating a PR. Often I'll also want to check the status of an
existing PR while working on that branch. To make both of those use cases more
efficient this change adds a `git pr` command that checks to see if a PR for
this branch exists. If so the PR is opened with the default browser, if not the
branch is automatically pushed up and a PR creation page is opened in the
default browser.

closes #6