[deduction] Improve interface for solving linear equations
In prototyping [pivot](https://github.com/caervs/rasm) I've noticed to
usability issues with pivot:
* The builtin row reduction for solving linear systems doesn't work in certain
  hard to reproduce situations
* There are occassions where a user of pivot may want some crude information
  about the system they are solving

To address the row reduction issue this change adds in the option to solve
systems using numpy instead of the builtin reduction and makes that option the
default until the builtin method has been thoroughly debugged. To address the
second concern users can now get the `variables` attribute of an expression,
equation, or equation set to get all variables that are present in the object.
This will help the pivot writer determine what anchor points exist in a drawing.