[xonsh] Ability to navigate with a side-bar
Use tmux and nnn to allow for quick navigation to other directories and files.
Invoking the navigation procedure will create an nnn sidebar with tmux and
monitor attempts to edit files. Meanwhile xonsh itself will follow those
directories and invoke the editor if the user tries to edit a file.