[bin] Tool for switching contexts to a given github issue
I've started tracking intentions for the day in emacs SES mode with the hope of
writing lots keyboard shortcuts to ease context switching. This tool lets me
have a dedicated tmux window for each project and with a single command switch
to the right window and branch for a given github issue.