[jql] Implement minor search
"Minor" search is search for entries in the current primary column matching the
search query, useful when looking for a known item or set of items in the
current table. A next step would be "major" search (hot key "?") to search for
all entries with any column matching the query.