[jql] Go to first foreign key
With more hierarchical table structures (each entry has a well defined parent
and 0 to many well-defined children), it can be cumbersome to have to navigate
to the parent entry in the table to go to the parent. Switch the 'go-to'
behavior to look for the first entry in the current row (starting at the
primary selection and wrapping around) that is a foreign key.