Add a macro for automatic attention cycle generation
Part of the timedb schema, this will allow you to select an item you would like
to be in your attention for a given cycle (quarter) and it will auto-create an
attention task for it. The macro goes further to group related tasks under
meaningful ancestors that finds a balance between decluttering and not
introducing too many levels of indirection.

Create a jql python module to contain the newly common function for generating task pks