Manually opt attention cycles out of jql execution view
Attention cycles are one example of a task we don't want to be distracted by
in our daily execution/day planning. In fact this rule applies to any task
which we want to periodically evaluate to see how we're doing against it or
whose work has already been fully broken down (a form of periodic evaluation).

How we'll achieve this instead will be to explicitly mark such tasks as
"Habitual" and review them as a group with a new execution view.