Show attention cycles in jql execution view
Use the 'a' key to switch modes between a view that will list all your
active/pending tasks and a view that will show the attention cycles to which
they belong. This view gives the user a sense of what are the higher-level
domains to which (s)he is attending and if each is getting adequate care. It
also allows the user to hop to a higher-level task (e.g. for time tracking that
doesn't benefit from being so precise).