Distinguish main view mode from task view mode in jql executor
Main view mode has come to be used for transitioning between various states
whereas task view mode is taking on the original purpose here which was to
distinguish how tasks are presented. This fixes a bug when trying to go to the
JQL entry for a task in attention view mode.