Add an actions table to jql timedb
Previous behavior was to have a hard-coded list of actions when setting PKs
with implicit types/semantics for their direct/indirect objects. This change
makes that more explicit and more flexible with an "actions" table defined in
the timedb that allows you to automatically populate attributes for the task
and set its pk.